requirements_latest.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_anaconda_depends=2019.03=py37_0
_libgcc_mutex=0.1=main
absl-py
alabaster
anaconda
anaconda-client
anaconda-project
appdirs
asn1crypto
astor
astroid
astropy
atomicwrites
attrs
audioread
autopep8
babel
backcall
backports
backports.os
backports.shutil_get_terminal_size
beautifulsoup4
bitarray
bkcharts
black
blas
bleach
blosc
bokeh
boto
boto3
botocore
bottleneck
bzip2
ca-certificates
cachetools
cairo
captum
certifi
cffi
chardet
click
cloudpickle
clyent
colorama
contextlib2
cryptography
cudatoolkit
cudnn
cupy
curl
cursor
cycler
cython
cytoolz
dask
dask-core
dbus
decorator
defusedxml
deprecation
distributed
docutils
emoji
entrypoints
et_xmlfile
expat
fastcache
fastrlock
filelock
filetype
flask
fontconfig
freetype
fribidi
fsspec
future
gast
get_terminal_size
gevent
glib
glob2
gmp
gmpy2
google-auth
google-auth-oauthlib
google-pasta
graphite2
greenlet
grpcio
gst-plugins-base
gstreamer
h5py
halo
harfbuzz
hdf5
heapdict
html5lib
icu
idna
imageio
imageio-ffmpeg
imagesize
imbalanced-learn
importlib_metadata
intel-openmp
ipykernel
ipython
ipython_genutils
ipywidgets
isort
itsdangerous
jbig
jdcal
jedi
jeepney
jinja2
jmespath
joblib
jpeg
json5
jsonschema
jupyter
jupyter_client
jupyter_console
jupyter_core
jupyterlab
jupyterlab_server
keras
keras-applications
keras-preprocessing
keyring
kiwisolver
krb5
lazy-object-proxy
libarchive
libcurl
libedit
libffi
libgcc-ng
libgfortran-ng
liblief
libpng
librosa
libsodium
libssh2
libstdcxx-ng
libtiff
libtool
libuuid
libxcb
libxml2
libxslt
lightgbm
llvmlite
locket
lxml
lz4-c
lzo
markdown
markupsafe
matplotlib
matplotlib-base
mccabe
mistune
mkl
mkl-service
mkl_fft
mkl_random
mock
more-itertools
moviepy
mpc
mpfr
mpmath
msgpack-python
multipledispatch
nbconvert
nbformat
nccl
ncurses
networkx
ninja
nltk
nose
notebook
numba
numexpr
numpy
numpy-base
numpydoc
oauthlib
olefile
openpyxl
openssl
opt-einsum
packaging
pandas
pandoc
pandocfilters
pango
parso
partd
patchelf
path.py
pathlib2
pathspec
patsy
pcre
pep8
pexpect
pickleshare
pillow
pip
pixman
pkginfo
plotly
pluggy
ply
proglog
prometheus_client
prompt_toolkit
protobuf
psutil
ptyprocess
py
py-lief
pyasn1
pyasn1-modules
pycodestyle
pycosat
pycparser
pycrypto
pycurl
pydub
pyflakes
pygments
pylint
pyodbc
pyopenssl
pyparsing
pyqt
pyrsistent
pysocks
pysoundfile
pytables
pytest
pytest-arraydiff
pytest-astropy
pytest-doctestplus
pytest-openfiles
pytest-remotedata
python
python-dateutil
python-libarchive-c
pytorch
pytz
pywavelets
pyyaml
pyzmq
qt
qtawesome
qtconsole
qtpy
readline
regex
requests
requests-oauthlib
resampy
retrying
ripgrep
rope
rsa
ruamel_yaml
s3transfer
sacremoses
scikit-image
scikit-learn
scipy
seaborn
secretstorage
send2trash
sentencepiece
setuptools
simplegeneric
singledispatch
sip
six
snappy
snowballstemmer
sortedcollections
sortedcontainers
soupsieve
speechrecognition
sphinx
sphinxcontrib
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sphinxcontrib-websupport
spinners
spyder
spyder-kernels
sqlalchemy
sqlite
statsmodels
sympy
tbb
tblib
tensorboard
tensorboardx
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
text-unidecode
tk
tokenizers
toml
toolz
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
typed-ast
typing
umap-learn
unicodecsv
unixodbc
urllib3
wandb
wcwidth
webencodings
werkzeug
wheel
widgetsnbextension
wrapt
wurlitzer
xgboost
xlrd
xlsxwriter
xlwt
xz
yaml
zeromq
zict
zipp
zlib
zstd
